Switch to libuv to manage subprocess lifetime.
In the not so far future, we will be removing support for making use of
process descriptors directly. Subprocesses may only be spawned through
an event loop, which will be libuv compatible in our case. Prepare for
the removal of the process descriptor functions by switching to libuv
now.

As our copy of GRPC doesn't make use of libuv (yet?), let's extend this
class to allocate a single event loop for all instances. That way we
only need to care about managing a single handle, as opposed to
maintaining a full loop for every instance.